pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "The program determines difference of two files (support JSON and YAML files)"
authors = ["Sergey Kornienko <skornienko29@gmail.com>"]
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.8"
pytest-cov = "^2.11.1"
pytest = "^6.2.1"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"
